Don't allow cloning HPKE private keys
This is the first step towards getting rid of HPKE key definitions in the ciphersuite. This removes the ability to clone HPKE private keys, which requires some changes to how state is moved around.
I didn't add any tests here but the paths are tests with the existing group tests and additional tests should be written for the ciphersuite and relevant paths later.