unified error handling 1/2
This is the first PR for unifying and cleaning up the error handling.
Each module now has a `errors.rs` that defines errors for this module.
Only errors from this file are allowed to be used in functions of this
module. The `errors.rs` takes care of converting other errors to errors
of the respective module.

Because this is already pretty big I stopped to split this in two.
Not all modules are converted yet.