Add encoding/decoding tests for MLSPlaintexts, MLSCiphertexts and Welcome messages (#203)
This PR adds a number of encoding/decoding tests for anything that goes over the wire. In the process of writing the tests, a few fixes were made in decoding functions, as well as an overhaul of extensions handling in the KeyPackageBundle constructor, which now makes sure that the input `ciphersuites` and the `CapabilitiesExtension` match and that the mandatory extensions (i.e. a `CapabilitiesExtension`, as well as a `LifetimeExtension`) are present, or, if not, that they are added with default setting. One TODO that remains is setting a default lifetime for the `LifetimeExtension`, which should be done as part of #164 .