Fuzzing targets for MLSCiphertext and MLSPlaintext decode (#212), as well as a few bug fixes
Introduce new fuzzing targets for the `decode` functions of MLSCiphertext and MLSPlaintext and integrates them into the CI. Thus fixing #154.

This commit also fixes a few issues uncovered during initial fuzzing, such as missing decode functions, which were previously removed as part of #203. It also fixes a bug, where the `RatchetTreeExtension` is removed from the Welcome message prior to verifying its signature, which lead to the verification failing reliably if a `RatchetTreeExtension` is present.