Key Schedule Redesign (#225)
This commit refactors the schedule module to be a bit more robust. Each individual Secret, i.e. the InitSecret, JoinerSecret, EpochSecret, etc. gets their own type such that we can guarantee that they are created and used in the way they are meant to be used in the Spec, creating what is essentially a state-machine with well-define transitions. Secrets are consumed as soon as they are not needed anymore to enable us to achieve FS. The design is not entirely "tight", as we still need to clone InitSecrets. Ideally, we would also consume the CommitSecret, but before we can do that, a refactoring (with FS in mind) of the PrivateTree is necessary (#224).

This commit also touches a few other places in the implementation to make them work with the new key schedule or to make the consistent with the design.