Errors for OpenMLS (#254)
This PR changes the way errors are defined and handled. I wasn't
really happy with the way this worked before.

There are two different types of errors allowed in OpenMLS now:
* Simple errors
  enum MyErrorType {
    ErrorVariant = "description",
  }
* Complex errors
  enum MyOtherErrorType {
    ErrorVariant(MoreDetaildError) = "description",
  }
Every definition of an error must be defined in the implement_error
macro, which ensures that conversion and all necessary functions
for the error are defined.

The top level error.rs also provides to basic error types for
complex errors that are not so complex.
ErrorString(String) and ErrorPayload(Vec<u8>).

The errors for the group and managed group are changed to work in
the new way. All new errors should be defined in this way and a
follow up PR will change all other existing errors to follow this
pattern.