Test vector generation and testing - TreeMath and Encryption (#299)
This implements test vector generation and testing for treemath
and encryption. It follows the basic idea of
https://github.com/mlswg/mls-implementations/blob/87e47925116b95c01db0e7b2a03b71b5a44b8a30/test-vectors.md
but slightly changes the test vector definition and uses JSON
serialised representations.
This version should be used to discuss the exact test vectors we
want for MLS and will probably change later on.

Also see https://github.com/mlswg/mls-implementations/issues/32
for a discussion.

Note that generating and testing for the encryption is currently
pretty cumbersome and ugly due to OpenMLS' code structure. This
should change in future when the code is refactored to respect
logic boundaries between the different parts of MLS.