key schedule reorg (#330)
In preparation for the key schedule test vectors I re-organised the
key schedule and moved the logic into the actual key schedule module
instead of the group. 
The changes are summarised in the module description of the schedule
module.

There are no changes to functionality PR, only moving code around.