Make Secret protocol version and cipher suite aware
The main purpose of this patch is to make the labeled key derivation
dependent on the cipher suite and MLS version (#231).

This patch stores the the cipher suite and MLS version of a secret
as part of the struct and enforces that it is only used with compatible
other secrets. It also removes the need to pass the ciphersuite to
many functions.
At the same time this increases the complexity of creating and parsing
secrets and other structs that don't have the necessary information.

This patch also introduces constant time comparison for secrets, in
particular for Mac values, which are now secrets as well.

Fixes #231