Change the Managed Group to accept a KeyStore ref instead of keeping a CredentialBundle ref (#385)
This commit adds a `key_store` module to OpenMLS. The `key_store` module provides a simple `KeyStore` struct, the purpose of which is to manage key material that is used in the individual groups. Initially, it only manages the "init" key packages and credentials, as well as the corresponding private keys.