use tls-codec for tree hash structs (#394)
This PR starts using the tls-codec for OpenMLS TLS encoding, starting
with the tree hash structs. Note that this is a mix now, but the tree
hash inputs are pretty self contained and therefore a good starting
point. Some of the encoding is a little more manual than before. This
should be improved later on when more structs use the tls-codec crate.

The tls-codec crate now offers a way to pre-allocate memory for
serializing. The `TlsSize` trait must be implemented for any struct
that want to implement `Serialize` and provides a `serialized_len`
method to get the length of the struct before serializing.

This reduces the tree_hash time by about 30%. There's a lot more to
do here, but this is an easy start to improve performance.