API changes and fixes for TreeKEM Test Vector integration (#390)
This commit changes the default behaviour for the `add_members` function in the Managed API such that it always forces a "self update", i.e. that the commit includes a path.

It also includes a few fixes to the code that surfaced due to the forced path inclusion.

One of the fixes in this PR fixes #380.

This commit additionally extends the Managed API slightly with a few functions that are only exposed for tests or interop testing.