Fix bug in path_secrets function and include TreeKEM test vector (#399)
The final commit in a sequence of commits to integrate message and TreeKEM test vectors. As with the previous commits, this is mostly @franziskuskiefer's code that I'm organizing into commits.

This commit notably includes a fix to the `path_secret` function, which now checks if nodes are blank before assuming that we have a given path secret.