signature refactor (#418)
In preparation for the new key store I cleaned up the way signing and verification with credentials works.
With this PR the signable module is the only place where sign and verify on credentials is called. This will make it easier to move around credentials and in particular the underlying keys.

To achieve this signed structs needed a little refactor. Notably, KeyPackage/KeyPackageBundle can't be modified anymore but modifying requires to convert them to unsigned representations. Signing gets you back the KeyPackage/KeyPackageBundle.

Another notable and somewhat unrelated change in this PR is that comparing KeyPackage now ignores the signature. This is necessary because the key package might be signed twice such that a different signature is attached even though the key package is the same. This is probably the better way to define PartialEq for key packages and necessary for KAT tests that re-sign key packages.

There will be a follow up to treat MlsPlaintext similarly to the KeyPackage making the signature a mandatory field rather than something that's added later.