rename add only commit to partial commit (#446)
We don't have the notion of an explicit "add only" commit. The mechanics are there, though. So on the level of the MLS Group API, one can not have any remove/update propsals and don't flip the "force self update" bit, which will result in an "add only"/"partial" commit.