Enable Managed API to process VerifiableMlsPlaintext (#444)
This commit enables the Managed API to process VerfiableMlsPlaintext objects, which is required, as it's not possible (and not meant to be) to deserialize `MlsPlaintext` directly. This ocmmit creates a new enum `MlsMessageIn` which contains either an unverified plaintext or a ciphertext. The ManageAPI then decrypts the ciphertext (decryption already deals with verification of the contained ciphertext) or verifies the plaintext (this step was missing) and continues processing.

This is a temporary fix until verification is dealt with properly as per #133.

Co-authored-by: Franziskus Kiefer <franziskuskiefer@gmail.com>