Implement type-enforced verification for `PublicGroupState` (#510)
As a preparation for the external init funcitonality, this commit implements the type-enforced verification pattern already in use for `MlsPlaintext` for `PublicGroupState`. See [the wiki](https://github.com/openmls/openmls/wiki/Signable) for more details. To de-clutter `messages/mod.rs` slightly, the PGS-related code is moved into its own file.

While implementing this, I noticed that `PublicGroupStateTbs` does not contain the group's ciphersuite, which I believe is an error in the spec (https://github.com/mlswg/mls-protocol/issues/498). This commit pre-emptively fixes this.