TreeSync integration (#611)
This commit removes large parts of the old `tree` module (keeping mostly `secrettree` and its dependencies) and instead integrates TreeSync/TreeKEM into the code.

In particular, it introduces the `treesync` module, which contains the `TreeSync` struct, as well as submodules for the diff functionality, TreeKEM, i.e. path encryption and decryption, as well as some submodules for node structs.

This commit touches a lot of the tests, since many tests directly access the tree of a given group, mostly to export nodes.

There is still `treemath` stub in the old `tree` module that contains the functions required for `secret_tree`.

Follow up issues are tracked in #423.