Use KeyPackageRef
Apologies for this humongous PR but I'm afraid there's no other way.
This PR uses the `KeyPackageRef` and `ProposalRef` values instead of leaf indices
and the proposal hash.

Instead of leaf indices the `KeyPackageRef` value (16 bytes truncated hash of
the key package) is used to identify leaves now.
There should be no functional changes in this PR.

I filed a number of follow ups that are necessary (labeled with KeyPackageRef).
It's mostly tests that are either missing or don't work anymore the way they used
to.

Note that the secret tree still uses leaf indices.

Fixes 541
Fixes 537