Partial support for external init (#634)
This commit adds partial support for the external init operation, where a non-member can inject themselves into a group. In particular, this commit allows the creation of a group from a `PublicGroupState` object and the creation of a corresponding external commit. It also enables other group members to process this commit. However, it does not yet allow the externally committing group member to process their own commit. The overall effort of enabling external init in OpenMLS is tracked in #630.

Concretely, this commit introduces a new `new_from_external_init` function to the `MlsGroup` API, which results in a new group and the commit message. The other main changes are to `create_commit` and `stage_commit` to enable commits by non-members (i.e. `SenderType::NewMember`) and their processing by group members. Changes in other places should all be related to the changes in those functions. In addition, this PR also includes a test that mostly serves to show that the basic functionality is ensured. 

Co-authored-by: Franziskus Kiefer <franziskuskiefer@gmail.com>
Co-authored-by: raphaelrobert <raphaelrobert@users.noreply.github.com>