Refactor proposal queues (#667)
This commit refactors proposal queues and in the process
 - gets rid of the `CreationProposalQueue` and the associated implementations and structs
 - gets rid of the `apply_proposal` function (thus fixing #565)
 - Removes the `Staged...` prefix whenever it's followed by `Proposal...`, as these structs are now used both for creation and staging

The advantage of the `CreationProposalQueue` was that it relied on references rather than full `Proposal` structs. However, in the course of #617 we started returning a `StagedProposalQueue` at the end of `create_commit`, so the cloning was taking place in any case.