Move message_secrets_store into CoreGroup
In order to use KeyPackageRefs we need to restructure the way the message
decryption works. This PR is a step in this direction by moving the
message_secrets_store into the CoreGroup and having one way of getting it.
This is not guaranteed to fully make the KeyPackageRef work but is a step in
this direction.

Other changes:
Derypting ciphertext is in two steps now, sender data and actual ciphertext.
This will make it again easier to use KeyPackageRef because the two
decryption steps (header and payload) are separated.