Expose list of member KeyPackages in public API (#747)
This commit changes the existing `members` function to return a vector of KeyPackages instead of credentials. This allows consumers to examine the corresponding credentials or compute KeyPackageRefs for use with other API endpoints. It also adds some documentation around this API endpoint to the book and extends some tests to make use of the new functionality.

Fixes #738.