Have removed members merge proposal changes to their TreeSync instance (#757)
This commit changes the bevahiour of `staged_commit` in `core_group` such that it does not immediately return after it notices that "self" was removed from the group. Instead, it includes a `StagedTreeSyncDiff` in the returned `StagedCommit`, which includes the results of all changes made in `apply_proposals`. The `StagedCommit` can then be merged after inspection, such that the membership list (but not the whole group state) of the removed client is the same as that of other clients after processing a given commit.

This commit also adds a test to test this new behaviour.

Fixes #616.

Co-authored-by: raphaelrobert <raphaelrobert@users.noreply.github.com>