Remove Send + Sync requirement on OpenMlsKeyStore
Send + Sync requirement is needed for rayon to work. It turns out that
OpenMlsKeyStore is never accessed in parallel so such bound is not
necessary.

Use case: rusqlite's Connection is not thread-safe so it is not possible
to save credential and key package bundles in SQLite out of the box.

Possible alternatives:

1. Unsafely implement Send + Sync on Connection.
    This is what Wire app uses ([Reference](https://github.com/wireapp/core-crypto/blob/develop/keystore/src/connection/platform/generic/mod.rs#L29))
2. Add a cargo feature to disable parallelization.

What do you think?