Drop key package bundle (#1160)
KeyPackageBundle is mostly gone and only used internally as a shallow wrapper to pass around KeyPackages together with their private keys. It is also used in tests for simplicity but could be removed there when rewriting tests.
Instead, the KeyPackage is used everywhere now and the corresponding private key is stored in the key store.

This also introduces a CryptoConfig to pass around the Ciphersuite and ProtocolVersion together.