Pull own_index out of TreeSync (#1216)
This PR pulls the `own_index` out of TreeSync and TreeSyncDiff and instead stores it in CoreGroup. Functions in TreeSync that rely on an `own_index` now get it as an additional input.

One thing I noticed during testing is that indexing encryption keypairs in the KeyStore with the `identity` (which is what was introduced in #1203) is not great, because MLS doesn't really have a notion of `identity` anyway and instead relies completely on the application to keep track of clients' identities. This PR changes the indexing in the key store such that instead of the `identity`, it uses the client's leaf index in the group. The index is thus a combination of group ID, group epoch and leaf index, which should uniquely identify a given group member and allow storage of key material for multiple members of the same group (which is mostly relevant for testing).

This PR is again in preparation of having a notion of a public group that can be used by a non-group member (such as a DS or even an external committer).