Introduce PublicGroup (#1218)
This PR introduces the notion of a `PublicGroup` and is follow-up based on #1216. The PublicGroup contains only public values associated with an MLS group and lives in a CoreGroup. Long-term, the goals is to provide an API (maybe under a feature flag) that allows public observers of MLS public messages (such as a delivery service) to use the public group to keep track of the group and its membership.

Concretely, this PR does the following:
* Introduce PublicGroup within CoreGroup (including PublicGroupDiffs).
* Move most parts of CoreGroup that deal with public group values into PublicGroup. For size reasons, this PR doesn't move all such values and functions into PublicGroup. In particular, validation is still part of CoreGroup. More in the next follow-up.
* Hide the tree from CoreGroup to a large extent. Again for size reasons, it doesn't hide it full yet (more in a separate follow-up). There are still myriad places where the treesync is exposed (e.g. via the `treesync` getter of CoreGroup), so this would add quite a few more changes to this PR.
* Fixes a small bug in the tree logic (lookup of nodes in a diff) and adds a test.
* Does some restructuring in create_commit and stage_commit. These are simplifications that are possible due to the recent PRs and overall improvements to OpenMLS :-)
* Restructures how and when the context is updated s.t. it fixes #931.