Add credentials_to_verify method to  StagedCommit
A StagedCommit can carry credentials in several different places,
all of which have to be verified. To make this easier on the caller,
we add the `credentials_to_verify` method on StagedCommit, which returns
an iterator over all the credentials in a StagedCommit that need to
be verified.

Also we add infrastructure in test_framework for letting the test code
decide which credentials would pass validation, by passing in a closure
that takes a &Credential and returns a bool.