Refactor VerifiedStruct and Verifiable (#1472)
The current version of the two traits is the way it is because it attempted
to prevent the caller from being able to convert to the Type implementing
VerifiedStruct without checking the signature. However, the approach that
was taken does not work.

This commit refactors the two traits and performs the conversion in the
same method as the verification. Unfortunately, this means that the
verify method has to be defined per impl instead of being able to rely
on a default one. There probably is no straightforward way around that,
so this simple solution is preferred.


Co-authored-by: Konrad Kohbrok <konrad.kohbrok@datashrine.de>