Fix validation bug (#1491)
This PR addresses issue #1487. The problem is that currently the list of extensions in a GroupContextExtensions proposal is compared against the capabilities of the current members. Instead, it needs to

 -   check that the RequiredCapabilities of the group are satisfied by all members of the group
 -   check that the types of all extensions in the proposal are in the RequiredCapabilities.

This PR also adds a method on the RequiredCapabilitiesExtension for checking whether an extension type is supported. This is needed because the extension types defined in the RFC are supported implicitly, so we also have to compare against the defaults. The defaults are defined close to the Capabilities struct and used to be private to these modules. I made the one for extension types pub(crate) so I can check against it.

---------

Co-authored-by: Jan Winkelmann (keks) <jan@cryspen.com>
Co-authored-by: raphaelrobert <raphaelrobert@users.noreply.github.com>