Make StagedWelcome consume Welcome instead of MlsMessageIn (#1533)
In #1506 we not only added the StagedWelcome types, but also made the methods to create them consume an MlsMessageIn instead of a Welcome. This was done at the time because there was the open issue #1326 that asked to do this, and we didn't check whether that really was a good idea.

It wasn't.

This PR keeps StagedWelcome, but makes the method that create these structs a Welcome instead of a MlsMessageIn.

Fixes #1519.