chore(ephemeral): make `KeyPackageSecretEncapsulation` public, serializable
CoreCrypto needs the ability to export and import KeyPackages for the
ephemeral client: the history secret has to include the KeyPackage and
all relevant supporting data, and the ephemeral client needs to be able
to restore it all in order to initialize itself.

Keypackages are part of a triad of related encryption data, which
is collected in this struct. But that struct was previously not public.

This commit makes it public in preparation for export and import.