chore: No external pub extension in welcome group info (#1851)
We currently put an external_pub GroupInfo extension into the GroupInfo that ends up in a Welcome. That's unnecessary and bloats the Welcome, especially when using a PQ Ciphersuite.

This PR changes the commit builder s.t. GroupInfos are generated separately for Welcomes and as a return value for the builder. This can result in an extra signature computation, but allows us to keep the GroupInfo in the Welcome clean of unnecessary extensions.